Fix: Security 순환 참조 문제 해결
