+++ src/main/java/com/the_dapda/domain/user/repository/UserRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-// 예제 13.7
+
 public interface UserRepository extends JpaRepository<User, Long> {
 
     Optional<User> findById(String id);
