--- src/main/java/com/the_dapda/domain/user/service/MemberService.java
@@ -1,10 +1,0 @@
-package com.the_dapda.domain.user.service;
-
-import com.the_dapda.domain.user.dto.UserDto;
-import com.the_dapda.global.response.ResponseForm;
-
-public interface MemberService {
-    ResponseForm register(UserDto memberDto);
-    ResponseForm login(UserDto memberDto); // 로그인 기능 추가
-    ResponseForm logout();
-}
