+++ src/main/java/com/the_dapda/global/security/config/CustomAuthenticationEntryPoint.java
@@ -13,10 +13,6 @@ import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.stereotype.Component;
 
-/**
- * 인증 실패시 결과를 처리해주는 로직을 가지고 있는 클래스
- * 예제 13.21, 예제 13.32
- */
 @Component
 public class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint {
