+++ src/main/java/com/the_dapda/global/security/config/PasswordEncoderConfiguration.java
@@ -5,7 +5,6 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.security.crypto.factory.PasswordEncoderFactories;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
-// 예제 13.26
 @Configuration
 public class PasswordEncoderConfiguration {
