+++ src/main/java/com/the_dapda/global/security/provider/JwtTokenProvider.java
@@ -1,11 +1,7 @@
 package com.the_dapda.global.security.provider;
 
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jws;
 import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
 import java.nio.charset.StandardCharsets;
-import java.util.Base64;
 import java.util.Date;
 import java.util.List;
 
@@ -24,16 +20,6 @@ import org.springframework.stereotype.Component;
 import javax.crypto.SecretKey;
 import javax.crypto.spec.SecretKeySpec;
 
-/**
- * JWT 토큰을 생성하고 유효성을 검증하는 컴포넌트 클래스 JWT 는 여러 암호화 알고리즘을 제공하고 알고리즘과 비밀키를 가지고 토큰을 생성
- * <p>
- * claim 정보에는 토큰에 부가적으로 정보를 추가할 수 있음 claim 정보에 회원을 구분할 수 있는 값을 세팅하였다가 토큰이 들어오면 해당 값으로 회원을 구분하여 리소스
- * 제공
- * <p>
- * JWT 토큰에 expire time을 설정할 수 있음
- */
-
-// secretKey : String 대신 SecretKey type으로 변경
 @Component
 @RequiredArgsConstructor
 public class JwtTokenProvider {
@@ -46,21 +32,16 @@ public class JwtTokenProvider {
     private final long tokenValidMillisecond = 1000L * 60 * 60; // 1시간 토큰 유효
     private SecretKey secretKey;
 
-    /**
-     * SecretKey 에 대해 인코딩 수행
-     * 예제 13.11
-     */
+
     @PostConstruct
     protected void init() {
         LOGGER.info("[init] JwtTokenProvider 내 secretKey 초기화 시작");
         System.out.println(secretKeyStr);
-//        secretKey = Base64.getEncoder().encodeToString(secretKey.getBytes(StandardCharsets.UTF_8));
         secretKey = new SecretKeySpec(secretKeyStr.getBytes(StandardCharsets.UTF_8), Jwts.SIG.HS256.key().build().getAlgorithm());
         System.out.println(secretKey);
         LOGGER.info("[init] JwtTokenProvider 내 secretKey 초기화 완료");
     }
 
-    // 예제 13.12
     // JWT 토큰 생성
     public String createToken(String userUid, List<String> roles) {
         LOGGER.info("[createToken] 토큰 생성 시작");
@@ -77,24 +58,19 @@ public class JwtTokenProvider {
         return token;
     }
 
-    // 예제 13.13
     // JWT 토큰으로 인증 정보 조회
     public Authentication getAuthentication(String token) {
         LOGGER.info("[getAuthentication] 토큰 인증 정보 조회 시작");
         UserDetails userDetails = userDetailsService.loadUserByUsername(this.getUsername(token));
         LOGGER.info("[getAuthentication] 토큰 인증 정보 조회 완료, UserDetails UserName : {}",
-            userDetails.getUsername());
+                userDetails.getUsername());
         return new UsernamePasswordAuthenticationToken(userDetails, "",
-            userDetails.getAuthorities());
+                userDetails.getAuthorities());
     }
 
-    // 예제 13.14
     // JWT 토큰에서 회원 구별 정보 추출
     public String getUsername(String token) {
         LOGGER.info("[getUsername] 토큰 기반 회원 구별 정보 추출");
-
-//        String info = Jwts.parser().verifyWith(secretKey).parseClaimsJws(token).getBody()
-//            .getSubject();
 
         String info = Jwts.parser()
                 .verifyWith(secretKey)
@@ -106,26 +82,15 @@ public class JwtTokenProvider {
         return info;
     }
 
-    // 예제 13.15
-    /**
-     * HTTP Request Header 에 설정된 토큰 값을 가져옴
-     *
-     * @param request Http Request Header
-     * @return String type Token 값
-     */
     public String resolveToken(HttpServletRequest request) {
         LOGGER.info("[resolveToken] HTTP 헤더에서 Token 값 추출");
         return request.getHeader("X-AUTH-TOKEN");
     }
 
-    // 예제 13.16
     // JWT 토큰의 유효성 + 만료일 체크
     public boolean validateToken(String token) {
         LOGGER.info("[validateToken] 토큰 유효 체크 시작");
         try {
-//            Jws<Claims> claims = Jwts.parser().setSigningKey(secretKey).parseClaimsJws(token);
-//            LOGGER.info("[validateToken] 토큰 유효 체크 완료");
-//            return !claims.getBody().getExpiration().before(new Date());
             return ! Jwts.parser()
                     .verifyWith(secretKey)
                     .build()
